+++ tests/TestUtils/ExtraAssertions/AssertReturnsCarbonNowTests.php
@@ -9,7 +9,7 @@ use PHPUnit\Framework\TestCase;
 use Raptor\TestUtils\ExtraAssertions;
 
 /**
- * Tests for _ExtraAssertions::assertReturnsCarbonNow_.
+ * Tests for _ExtraAssertions::assertReturnsCarbonNow_ and _ExtraAssertions::assertReturnsCarbonNowWithoutMicroseconds_.
  *
  * @author Mikhail Kamorin aka raptor_MVK
  *
@@ -59,6 +59,18 @@ final class AssertReturnsCarbonNowTests extends TestCase
      */
     public function assertReturnsCarbonNowRejectDataProvider(): array
     {
+        return $this->prepareWrongAssertReturnsCarbonNowTestData();
+    }
+
+    /**
+     * Prepares test data for _assertReturnsCarbonNow_ to reject.
+     *
+     * @return array [ [ func ], ... ]
+     *
+     * @SuppressWarnings(PHPMD.StaticAccess) __approved__ factory method Carbon::now
+     */
+    private function prepareWrongAssertReturnsCarbonNowTestData(): array
+    {
         $wrongCarbon = static function () {
             return Carbon::now()->addDay();
         };
@@ -73,5 +85,56 @@ final class AssertReturnsCarbonNowTests extends TestCase
             'not Carbon' => [$notCarbon, 'That is not Carbon'],
             'null Carbon' => [$nullCarbon, 'There is no Carbon']
         ];
+    }
+
+    /**
+     * Checks that assertion _assertReturnsCarbonNowWithoutMicroseconds_ accepts correct function.
+     *
+     * @SuppressWarnings(PHPMD.StaticAccess) __approved__ factory method Carbon::now
+     */
+    public function testAssertReturnsCarbonNowWithoutMicrosecondsAcceptsCorrectFunction(): void
+    {
+        $function = static function () {
+            return Carbon::now()->setMicro(0);
+        };
+
+        static::assertReturnsCarbonNowWithoutMicroseconds($function);
+    }
+
+    /**
+     * Checks that assertion _assertReturnsCarbonNowWithoutMicroseconds_ rejects incorrect function.
+     *
+     * @param callable $func
+     * @param string $message
+     *
+     * @dataProvider assertReturnsCarbonNowWithoutMicrosecondsRejectDataProvider
+     */
+    public function testAssertReturnsCarbonNowWithoutMicrosecondsRejectsIncorrectFunction(
+        callable $func,
+        string $message
+    ): void {
+        $messageRegExp = "/^$message\n.*$/";
+        $this->expectException(ExpectationFailedException::class);
+        $this->expectExceptionMessageRegExp($messageRegExp);
+
+        static::assertReturnsCarbonNowWithoutMicroseconds($func, $message);
+    }
+
+    /**
+     * Provides test data for _assertReturnsCarbonNowWithoutMicroseconds_ to reject.
+     *
+     * @return array [ [ func ], ... ]
+     *
+     * @SuppressWarnings(PHPMD.StaticAccess) __approved__ factory method Carbon::now
+     */
+    public function assertReturnsCarbonNowWithoutMicrosecondsRejectDataProvider(): array
+    {
+        $wrongCarbonMicroseconds = static function () {
+            return Carbon::now()->addMicrosecond();
+        };
+        return array_merge(
+            $this->prepareWrongAssertReturnsCarbonNowTestData(),
+            ['use microseconds' => [$wrongCarbonMicroseconds, 'Microseconds are used']]
+        );
     }
 }